+++ deci/src/jsMain/kotlin/org/kimplify/deci/Deci.kt
@@ -4,6 +4,7 @@ import kotlinx.serialization.Serializable
 import org.kimplify.deci.exception.DeciDivisionByZeroException
 import org.kimplify.deci.exception.DeciScaleException
 import org.kimplify.deci.parser.extractScale
+import org.kimplify.deci.parser.toDecimalLiteral
 import org.kimplify.deci.parser.validateAndNormalizeDecimalLiteral
 
 @Serializable(with = DeciSerializer::class)
@@ -18,7 +19,7 @@ actual class Deci private constructor(
 
     actual constructor(value: Long) : this(value.toString())
     actual constructor(value: Int) : this(value.toString())
-    actual constructor(value: Double) : this(value.toString())
+    actual constructor(value: Double) : this(value.toDecimalLiteral())
 
     actual companion object {
         actual val ZERO = Deci("0")